arch2_microservices/gateway_service/requirements.txt
Flask 
grpcio 
grpcio-tools